Various changes to make status bar and polling work in subscription, testrules and info page
