+++ src/main/resources/static/css/style.css
@@ -2,9 +2,6 @@
     padding: 10px;
 }
 .subscription-alert {
-    display: flex;
-    justify-content: center;
-    align-items: center;
     width: 100%;
     padding: 2px;
     margin-bottom: 0;
@@ -65,21 +62,6 @@
     width: 30%;
 }
 /* Information page stop */
-.testRulePage {
-    position: fixed;
-    top: 0px;
-    left: 0px;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, .5);
-    z-index: 999;
-    vertical-align: middle;
-    text-align: center;
-    color: white;
-    font-size: 25px;
-    font-weight: bold;
-    text-shadow: 1px 1px #525252;
-}
 .flexbox {
     display:flex;
     justify-content:center;
@@ -230,6 +212,13 @@ button.rules_info:focus {
     outline:0;
     outline:none;
 }
+button.remove-item {
+    padding: 0 2px 0 2px;
+}
+.text-area {
+    width: 100%;
+    padding-right: 5px;
+}
 #infoContent{
     font-family: Calibri;
 }
@@ -248,6 +237,9 @@ table.dataTable.dtr-inline.collapsed > tbody > tr[role="row"] > td:first-child:b
     left: 30%;
     border: none;
     box-shadow: 0 0 0 transparent;
+}
+table.dataTable.dtr-inline.collapsed > tbody > tr > td > .dtr-details > li {
+    word-break: break-word;
 }
 @media (max-width: 576px) {
     .container {
