+++ src/main/resources/static/css/ui.css
@@ -83,6 +83,7 @@ body {
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
+    word-break: break-word;
     padding: 10px 5px 10px 10px;
     font-size: 80%;
     cursor: pointer;
