+++ src/main/resources/static/js/subscription.js
@@ -4,7 +4,6 @@ var table;
 var frontendServiceUrl;
 var defaultFormKeyValuePair = { "formkey": "", "formvalue": "" };
 var defaultFormKeyValuePairAuth = { "formkey": "Authorization", "formvalue": "" };
-var timerInterval;
 
 jQuery(document).ready(function () {
 
@@ -92,9 +91,10 @@ jQuery(document).ready(function () {
     }
 
     // Check if EI Backend Server is online every X seconds
-    if (timerInterval == null) {
-        timerInterval = window.setInterval(function () { checkBackendStatus(); }, 15000);
+    if(timerInterval != null) {
+        window.clearInterval(timerInterval);
     }
+    timerInterval = window.setInterval(function () { checkBackendStatus(); }, 15000);
 
     // Check if buttons should be enabled or disabled
     // Toggle warning text on and off
