+++ src/main/resources/templates/subscription.html
@@ -13,7 +13,7 @@
             <div id="btnEIContainer">
                 <button
                     title="EI connection status check is refreshed continuously. Click button to force check status."
-                    type="button" id="btnEIConnection" class="btn btnEIConnectionStatus white-space-normal">
+                    type="button" id="btnEIConnection" class="btn btnEIConnectionStatus">
                     <i class="glyphicon"></i>EI Backend Status
                 </button>
             </div>
